Fix stupid crash when raft is generated as layer numbers can be negative.

+++ src/FffGcodeWriter.cpp
@@ -2759,8 +2759,8 @@ bool FffGcodeWriter::processSupportInfill(const SliceDataStorage& storage, Layer
             {
                 if (infill_extruder.settings.get<bool>("magic_spiralize"))
                 {
-                    size_t layer_nr = gcode_layer.getLayerNr();
-                    if(layer_nr > infill_extruder.settings.get<size_t>("bottom_layers"))
+                    int layer_nr = gcode_layer.getLayerNr();
+                    if(layer_nr > (int)infill_extruder.settings.get<size_t>("bottom_layers"))
                     {
                         // bit of subtlety here... support is being used on a spiralized model and to ensure the travel move from the end of the last spiral
                         // to the start of the support does not go through the model we have to tell the slicer what the current location of the nozzle is
